Add files via upload
For using BaCoCa open the terminal of your operating system. Move through your directory path to
the folder where BaCoCa is placed. To execute BaCoCa a PERL interpreter must be installed on the
current run system. Linux and Mac systems do normally not need a subsequent installation because
the PERL interpreter is a standard tool of these systems in advance, but have to install the additional

PERL Statistics::R package to use the ”-r” option of BaCoCa.vX.X.r.pl which allows the generation
of result heat maps by using R. For the usage of BaCoCa.vX.X.pl which has not included the ”-r” option,
additional PERL or R packages have not to be installed.